pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "emmentalembed"
version = "0.1.0"
description = "Protein embedding extraction and structure prediction at scale"
readme = "README.md"
requires-python = ">=3.11"
license = { text = "MIT" }
dependencies = [
    "numpy>=1.24.0",
    "pandas>=2.0.0",
    "biopython>=1.83",
    "pyyaml>=6.0",
    "typer>=0.12.0",
    "rich>=13.0.0",
    "tqdm>=4.66.0",
    "pyarrow>=15.0.0",
    "torch>=2.0.0",
    "transformers>=4.40.0",
    "tiktoken>=0.7.0",
    "sentencepiece>=0.2.0",
    "esm>=3.0.0",
    "chai_lab",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "ruff>=0.4.0",
]

[project.scripts]
emmentalembed = "emmentalembed.cli:app"

[tool.hatch.build.targets.wheel]
packages = ["src/emmentalembed"]

[tool.ruff]
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I", "D"]
pydocstyle = { convention = "google" }

[tool.ruff.lint.per-file-ignores]
"tests/*.py" = ["D100", "D103"]

[tool.pytest.ini_options]
testpaths = ["tests"]
